Rename flags file
