+++ packages/astro/src/cli/preview/index.ts
@@ -3,7 +3,7 @@ import type yargs from 'yargs-parser';
 import type { LogOptions } from '../../core/logger/core.js';
 import { printHelp } from '../../core/messages.js';
 import previewServer from '../../core/preview/index.js';
-import { flagsToAstroInlineConfig } from '../load-settings.js';
+import { flagsToAstroInlineConfig } from '../flags.js';
 
 interface PreviewOptions {
 	flags: yargs.Arguments;
