+++ packages/astro/src/cli/sync/index.ts
@@ -2,7 +2,7 @@ import type yargs from 'yargs-parser';
 import type { LogOptions } from '../../core/logger/core.js';
 import { printHelp } from '../../core/messages.js';
 import { sync as _sync } from '../../core/sync/index.js';
-import { flagsToAstroInlineConfig } from '../load-settings.js';
+import { flagsToAstroInlineConfig } from '../flags.js';
 
 interface SyncOptions {
 	flags: yargs.Arguments;
